feat: allow configuring of additional security group rules (#38)
## what

- This PR makes it possible to optionally add additional security group
rules to the main security group
- We add the following:
- A new tf resource `aws_security_group_rule.additonal`, which loops
over...
  - A new tf variable `var.additional_security_group_rules` of type map
  - Additions to the README.md

## why

- By allowing the configuration of sg rules directly in this module, we
do not require the module user to create additional security groups
outside this module. This is especially useful for those users that
consume this module with terragrunt, who may not have the ability to
easily create additional security groups

## references

- N/A, but as a user of downstream module
[terraform-aws-tailscale](https://github.com/masterpointio/terraform-aws-tailscale),
I would be delighted with this addition. If accepted, I will follow-up
with a PR in that module as well


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Added support for defining additional security group rules via a new
input variable.
- **Documentation**
- Updated documentation to include details about the new resource and
input variable for additional security group rules.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->