feat: allow additional custom IAM policy to attached EC2 role (#52)
## what

- Add `resource "aws_iam_role_policy" "custom"` to allow for a custom
policy document to be attached to the role that is associated with the
EC2 instance.

## why

- There are times where commands will be ran in the EC2 instance as the
`ec2-user`, either manually or via the `user_data`. This additional IAM
policy document will allow that.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- New Features
- Optional custom permissions: You can now attach a custom inline policy
to the default role by supplying a JSON policy.
- Configurable policy name: Provide a custom name for the attached
policy when using a custom policy.
- Non-intrusive default: If no custom policy is provided, nothing
changes for existing setups.
- Flexible integration: Enables tailoring role permissions to specific
organizational or workload needs without altering core configuration.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->